feat: abstract addperson command into its own command
