+++ src/main/java/seedu/address/logic/commands/AddPersonCommand.java
@@ -1,0 +1,96 @@
+package seedu.address.logic.commands;
+
+import static java.util.Objects.requireNonNull;
+import static seedu.address.logic.parser.CliSyntax.PREFIX_ADDRESS;
+import static seedu.address.logic.parser.CliSyntax.PREFIX_EMAIL;
+import static seedu.address.logic.parser.CliSyntax.PREFIX_NAME;
+import static seedu.address.logic.parser.CliSyntax.PREFIX_PHONE;
+import static seedu.address.logic.parser.CliSyntax.PREFIX_TAG;
+
+import seedu.address.commons.util.ToStringBuilder;
+import seedu.address.logic.Messages;
+import seedu.address.model.Model;
+import seedu.address.model.person.Person;
+
+/**
+ * Adds a person to the address book.
+ */
+public abstract class AddPersonCommand extends AddCommand {
+
+    public static final String ENTITY_STRING = "person";
+
+    public static final String MESSAGE_USAGE = COMMAND_WORD + ": Adds a person to the address book. "
+            + "Parameters: "
+            + PREFIX_NAME + "NAME "
+            + PREFIX_PHONE + "PHONE "
+            + PREFIX_EMAIL + "EMAIL "
+            + PREFIX_ADDRESS + "ADDRESS "
+            + "[" + PREFIX_TAG + "TAG]...\n"
+            + "Example: " + COMMAND_WORD + " "
+            + PREFIX_NAME + "John Doe "
+            + PREFIX_PHONE + "98765432 "
+            + PREFIX_EMAIL + "johnd@example.com "
+            + PREFIX_ADDRESS + "311, Clementi Ave 2, #02-25 "
+            + PREFIX_TAG + "friends "
+            + PREFIX_TAG + "owesMoney";
+
+    public static final String MESSAGE_SUCCESS = "New person added: %1$s";
+    public static final String MESSAGE_DUPLICATE_PERSON = "This person already exists in the address book";
+
+    private final Person toAdd;
+
+    /**
+     * Creates an AddPersonCommand to add the specified {@code Person}
+     */
+    public AddPersonCommand(Person person) {
+        requireNonNull(person);
+        toAdd = person;
+    }
+
+    @Override
+    protected boolean alreadyExists(Model model) {
+        return model.hasPerson(toAdd);
+    }
+
+    @Override
+    protected void addEntity(Model model) {
+        model.addPerson(toAdd);
+    }
+
+    @Override
+    protected String getSuccessMessage() {
+        return MESSAGE_SUCCESS;
+    }
+
+    @Override
+    protected String getDuplicateEntityMessage() {
+        return MESSAGE_DUPLICATE_PERSON;
+    }
+
+    @Override
+    protected Object formatEntity() {
+        return Messages.formatPerson(toAdd);
+    }
+
+    @Override
+    public boolean equals(Object other) {
+        if (other == this) {
+            return true;
+        }
+
+        // instanceof handles nulls
+        if (!(other instanceof AddPersonCommand)) {
+            return false;
+        }
+
+        AddPersonCommand otherAddCommand = (AddPersonCommand) other;
+        return toAdd.equals(otherAddCommand.toAdd);
+    }
+
+    @Override
+    public String toString() {
+        return new ToStringBuilder(this)
+                .add("toAdd", toAdd)
+                .toString();
+    }
+}
